rating added

+++ lib/Models/customer_models/spare_parts_list_model/spare_parts_list_model.dart
@@ -98,7 +98,7 @@ class SparePartsList {
     vehicleModel: VehicleModel.fromMap(json["vehicleModel"]),
     inCart: json["inCart"],
     reviewCount: json["reviewCount"],
-    avgRate: json["avgRate"],
+    avgRate: json["avgRate"].toDouble(),
 
   );
